requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
